rename files + fix paths
